+++ onebuild/utils.py
@@ -1,15 +1,7 @@
 # !/usr/bin/env python
 
-from onebuild import __version__
-
 DASH = '-' * 50
 NEWLINE = "\n"
-
-
-class PredefinedActions:
-    VERSION = "ONEBUILD_VERSION"
-    HELP = "ONEBUILD_HELP"
-    LIST = "ONEBUILD_LIST"
 
 
 def sample_yaml_file():
@@ -19,13 +11,5 @@ def sample_yaml_file():
            "  - lint: ./gradlew spotlessApply"
 
 
-def print_help(parser):
-    parser.print_help()
-
-
 def config_string(project):
     return "" + DASH + NEWLINE + project.__str__() + NEWLINE + DASH
-
-
-def version_string():
-    return '1build {version} '.format(version=__version__)